Refactor and add error handling
